requirements.txt
Django==6.0.4
djangorestframework==3.17.1
psycopg2-binary==2.9.11
bcrypt==5.0.0
PyJWT==2.12.1
python-dotenv==1.2.2
